Refactor `connect delete` structures
