+++ src/commands/connect/delete.ts
@@ -1,0 +1,43 @@
+import chalk from "chalk";
+import dedent from "ts-dedent";
+import { ConnectOptions } from ".";
+import logger from "../../util/logger";
+import { indent } from "../../util/text";
+import { getComponentConfigFiles } from "./config";
+import { ConnectedBarrels } from "./interfaces/api";
+import { ConnectedComponentsService } from "./service";
+
+const service = new ConnectedComponentsService();
+
+const deleteConnectedBarrels = async (connectedBarrels: ConnectedBarrels[]): Promise<void> => {
+    logger.info("Deleting connected components from Zeplin…");
+
+    await service.deleteConnectedBarrels(connectedBarrels);
+
+    logger.info("🔥 Component connections successfully deleted from components in Zeplin.");
+};
+
+const getConnectedBarrels = async (options: ConnectDeleteOptions): Promise<ConnectedBarrels[]> => {
+    const { configFiles } = options;
+
+    const componentConfigFiles = await getComponentConfigFiles(configFiles, []);
+
+    return componentConfigFiles as ConnectedBarrels[];
+};
+
+export type ConnectDeleteOptions = Pick<ConnectOptions, "configFiles">;
+
+export async function connectDelete(options: ConnectDeleteOptions): Promise<void> {
+    try {
+        const barrels = await getConnectedBarrels({ configFiles: options.configFiles });
+
+        await deleteConnectedBarrels(barrels);
+    } catch (error) {
+        error.message = dedent`
+            ${chalk.bold`Deleting connected components from Zeplin components failed.`}
+
+            ${chalk.redBright(indent(error.message))}
+        `;
+        throw error;
+    }
+}
